requirements.txt
fastapi
pydantic
poe-api
python-dotenv
openai
uvicorn
tqdm